static and dynamic data pathways
updated the pathways for how dynamic and static data is pulled into the map.
The static data is pulled from database and then used to generate the markers.
The dynamic data is generated with an sql procedure which i've created in the database (it returns the most recent stats for a station).  The Data is only called when the station is clicked.  Is quite slow which i think is due to the SQL query.  can explore other ways of saving time